Сделал загрузку данных с сервера

+++ app/src/main/java/ru/sergey/testalpha/presentation/viewmodel/MainUiState.kt
@@ -3,7 +3,7 @@ package ru.sergey.testalpha.presentation.viewmodel
 import ru.sergey.data.models.BinInfo
 
 data class MainUiState(
-    val binNumber: Int = 0,
+    val binNumber: Long = 0L,
     val binInfo: BinInfo? = null,
     val status: Status = Status.Idle
 ) {
